Aggiornata autenticazione

+++ ws-server/app.js
@@ -3,32 +3,25 @@ const cors = require('cors');
 const mysql = require('mysql');
 const bodyParser = require('body-parser');
 const jwt = require('jsonwebtoken');
-
+const config = require('./config.js');
+const auth = require('./authservice.js');
 
 var service = express();
 
 service.use(cors());
 service.use(bodyParser.json());
-service.use(bodyParser.urlencoded({extended: false}));
+service.use(bodyParser.urlencoded({ extended: false }));
 
 service.get('/', (request, response) => {
     response.sendFile(__dirname + '/index.html');
 });
 
-const config = {
-    host: 'sql-progetto-prova',
-    user: 'utente',
-    password: 'root',
-    database: 'dati-progetto'
-};
-
-let passPhrase = 'Questa è la chiave usata per generare il token';
 
 service.post('/login', (request, response) => {
     let username = request.body.username;
     let password = request.body.password;
-    
-    let mysqlconn = mysql.createConnection(config);
+
+    let mysqlconn = mysql.createConnection(config.dbConfig);
 
     let sqlString = "SELECT * FROM Admins WHERE username = ? AND pwd = ?;";
     let params = [username, password];
@@ -37,8 +30,9 @@ service.post('/login', (request, response) => {
         if (!error) {
             if (data.length > 0) {
                 let token = jwt.sign({
-                    nome:data[0].nome,
-                    cognome: data[0].cognome}, passPhrase, {expiresIn: 60})
+                    nome: data[0].nome,
+                    cognome: data[0].cognome
+                }, config.passPhrase, { expiresIn: config.tokenDuration })
                 response.json(token);
             }
             else {
@@ -55,7 +49,7 @@ service.post('/login', (request, response) => {
 
 service.get('/users', (request, response) => {
     // Stabilire la connessione con mysql server
-    let mysqlconn = mysql.createConnection(config);
+    let mysqlconn = mysql.createConnection(config.dbConfig);
 
     // Definire l'istruzione SQL da eseguire
     let sqlString = 'SELECT * FROM Users';
@@ -74,7 +68,7 @@ service.get('/users', (request, response) => {
 service.get('/users/:id', (request, response) => {
     let id = request.params.id;
     // Stabilire la connessione con mysql server
-    let mysqlconn = mysql.createConnection(config);
+    let mysqlconn = mysql.createConnection(config.dbConfig);
 
     // Definire l'istruzione SQL da eseguire
     let sqlString = 'SELECT * FROM Users WHERE id = ?';
@@ -93,7 +87,7 @@ service.get('/users/:id', (request, response) => {
 service.get('/users/gender/:gender', (request, response) => {
     let gender = request.params.gender;
     // Stabilire la connessione con mysql server
-    let mysqlconn = mysql.createConnection(config);
+    let mysqlconn = mysql.createConnection(config.dbConfig);
 
     // Definire l'istruzione SQL da eseguire
     let sqlString = 'SELECT * FROM Users WHERE gender LIKE ?';
@@ -113,7 +107,7 @@ service.get('/users/firstname/:firstName/lastname/:lastName', (request, response
     let firstName = request.params.firstName;
     let lastName = request.params.lastName;
     // Stabilire la connessione con mysql server
-    let mysqlconn = mysql.createConnection(config);
+    let mysqlconn = mysql.createConnection(config.dbConfig);
 
     // Definire l'istruzione SQL da eseguire
     let sqlString = 'SELECT * FROM Users WHERE first_name LIKE ? AND last_name LIKE ?';
@@ -129,47 +123,34 @@ service.get('/users/firstname/:firstName/lastname/:lastName', (request, response
     })
 });
 
-service.delete('/users/:id', (request, response) => {
+service.delete('/users/:id', auth, (request, response) => {
     let id = request.params.id;
-    let token = request.headers['x-access-token'];
 
-    if (!token) {
-        response.status(403).send('No token provided.');
-    }
-    else {
-        jwt.verify(token, passPhrase, (error, decoded) => {
-            if (error) {
-                response.status(403).send('Token invalid.');
-            }
-            else {                
-                let mysqlconn = mysql.createConnection(config);
-                
-                // Definire l'istruzione SQL da eseguire
-                let sqlString = 'DELETE FROM Users WHERE id = ?';
-                
-                // Eseguire l'istruzione SQL sulla connessione
-                mysqlconn.query(sqlString, id, (error, data) => {
-                    if (!error) {
-                        response.json(data);
-                    }
-                    else {
-                        response.status(500).send(error);
-                    }                
-                })
-            }       
-        });
-    }
+    let mysqlconn = mysql.createConnection(config.dbConfig);
+
+    // Definire l'istruzione SQL da eseguire
+    let sqlString = 'DELETE FROM Users WHERE id = ?';
+
+    // Eseguire l'istruzione SQL sulla connessione
+    mysqlconn.query(sqlString, id, (error, data) => {
+        if (!error) {
+            response.json(data);
+        }
+        else {
+            response.status(500).send(error);
+        }
+    })
 });
 
-service.post('/users/:id', (request, response) => {
+service.post('/users/:id', auth, (request, response) => {
     let id = request.params.id;
     let dati = request.body;
 
-    let mysqlconn = mysql.createConnection(config);
+    let mysqlconn = mysql.createConnection(config.dbConfig);
 
     let params = [id, dati.first_name, dati.last_name, dati.email, dati.gender, dati.ip_address];
     let sqlString = "INSERT INTO Users VALUES(?, ?, ?, ?, ?, ?);";
-    
+
     mysqlconn.query(sqlString, params, (error, data) => {
         if (!error) {
             response.json(data);
@@ -184,11 +165,11 @@ service.put('/users/:id', (request, response) => {
     let id = request.params.id;
     let dati = request.body;
 
-    let mysqlconn = mysql.createConnection(config);
+    let mysqlconn = mysql.createConnection(config.dbConfig);
 
     let params = [dati.first_name, dati.last_name, dati.email, dati.gender, dati.ip_address, id];
     let sqlString = "UPDATE Users SET first_name = ?, last_name = ?, email = ?, gender = ?, ip_address = ? WHERE id = ?;";
-    
+
     mysqlconn.query(sqlString, params, (error, data) => {
         if (!error) {
             response.json(data);
@@ -202,7 +183,7 @@ service.put('/users/:id', (request, response) => {
 
 
 
-var server = service.listen(3000, () => {
+var server = service.listen(config.port, () => {
     console.log('Server in ascolto sulla porta 3000...');
 })
